fix to margin in layout, refactor routes layout structure and remove unused components
