--- src/modules/common/components/Layout/index.tsx
@@ -1,9 +1,0 @@
-import { FC } from "react";
-import { LayoutProps } from "antd";
-import { StyledLayout } from "./styled";
-
-const Layout: FC<LayoutProps> = ({ children, ...props }) => {
-  return <StyledLayout {...props}>{children}</StyledLayout>;
-};
-
-export default Layout;
