--- src/modules/common/components/Layout/styled.ts
@@ -1,6 +1,0 @@
-import { Layout } from "antd";
-import styled from "styled-components";
-
-export const StyledLayout = styled(Layout)`
-  min-height: 100vh;
-`;
