--- src/modules/common/components/NavBar/styled.ts
@@ -1,10 +1,0 @@
-import { Layout } from "antd";
-import styled from "styled-components";
-
-export const StyledHeader = styled(Layout.Header)`
-  z-index: 1;
-  height: 70px;
-  display: flex;
-  padding: 15px 33px;
-  filter: drop-shadow(0px 2px 5px rgba(0, 0, 0, 0.1));
-`;
