--- src/modules/common/components/SideBar/types.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export interface SideBarProps {
-  children?: ReactNode;
-}
