+++ src/modules/common/components/index.ts
@@ -1,4 +1,2 @@
-export { default as NavBar } from "./NavBar";
-export { default as SideBarContent } from "./SideBar";
-export { default as Layout } from "./Layout";
+export * from "./Form";
 export * from "./RoutesWithNotFound";
