+++ src/router/AppRouter.tsx
@@ -1,7 +1,7 @@
 import { Navigate, Route } from "react-router-dom";
 import { AuthRoutes, AuthRoutesList } from "@/modules/auth";
 import { RoutesWithNotFound } from "@/modules/common/components";
-import { AdminLayoutRoutes } from "@/modules/common/layout/AdminLayout/AdminLayoutRoutes";
+import { AdminRoutes } from "@/router/AdminRoutes";
 import { DashboardRoutesList } from "@/modules/dashboard/routes";
 
 export const AppRouter = () => {
@@ -12,7 +12,7 @@ export const AppRouter = () => {
         path="/"
         element={<Navigate to={`${DashboardRoutesList.DASHBOARD}`} />}
       />
-      <Route path="/*" element={<AdminLayoutRoutes />} />
+      <Route path="/*" element={<AdminRoutes />} />
     </RoutesWithNotFound>
   );
 };
